+++ example/lib/const/resource.dart
@@ -1,0 +1,22 @@
+/// Generate by [resource_generator](https://github.com/CaiJingLong/flutter_resource_generator) library.
+/// PLEASE DO NOT EDIT MANUALLY.
+// ignore_for_file: constant_identifier_names
+class R {
+  const R._();
+
+  /// ![preview](file:///Users/zad/Desktop/flutter_assets_generator/example/assets/address.png)
+  static const String assetsAddressPng = 'assets/address.png';
+
+  /// ![preview](file:///Users/zad/Desktop/flutter_assets_generator/example/assets/address@at.png)
+  static const String assetsAddressATAtPng = 'assets/address@at.png';
+
+  /// ![preview](file:///Users/zad/Desktop/flutter_assets_generator/example/assets/bluetoothon-fjdfj.png)
+  static const String assetsBluetoothonFjdfjPng =
+      'assets/bluetoothon-fjdfj.png';
+
+  /// ![preview](file:///Users/zad/Desktop/flutter_assets_generator/example/assets/bluetoothon.png)
+  static const String assetsBluetoothonPng = 'assets/bluetoothon.png';
+
+  /// ![preview](file:///Users/zad/Desktop/flutter_assets_generator/example/assets/camera.png)
+  static const String assetsCameraPng = 'assets/camera.png';
+}