+++ lib/const/resource.dart
@@ -1,4 +1,5 @@
-/// Generate by [resource_generator](https://github.com/CaiJingLong/flutter_resource_generator) library.
+/// Generate by [flutter_assets_generator](https://github.com/goodswifter/flutter_assets_generator) library.
 /// PLEASE DO NOT EDIT MANUALLY.
 class R {
+  const R._();
 }